moved deprecated code renamed classes added flags to Descriptor added MultifeatureDescriptor changed point from int to double

+++ src/de/lmu/dbs/ImagePoint.java
@@ -2,24 +2,27 @@ package de.lmu.dbs;
 
 /*
  * A pixel representing a location in (x, y) coordinate plane, 
- * specified in integer precision.
+ * specified in double precision.
+ * 
+ * While images are usually only represented as ints, several algorithms use 
+ * sub pixel locations, so that doubles are more useful.
  */
 public class ImagePoint {
 
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
     // constructs and initializes a point at the specified location in the coordinate plane
-    public ImagePoint(int x, int y) {
+    public ImagePoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
